connect and detect working with fragments
