+++ app/src/main/java/com/specknet/pdiotapp/detect/DetectFragment.kt
@@ -1,9 +1,14 @@
 package com.specknet.pdiotapp.detect
 
 import android.content.BroadcastReceiver
+import android.content.Context
+import android.content.Intent
 import android.content.IntentFilter
 import android.os.Bundle
+import android.os.Handler
+import android.os.HandlerThread
 import android.os.Looper
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -21,11 +26,8 @@ import com.github.mikephil.charting.interfaces.datasets.ILineDataSet
 import com.google.android.gms.internal.zzhu.runOnUiThread
 import com.specknet.pdiotapp.R
 import com.specknet.pdiotapp.utils.Constants
+import com.specknet.pdiotapp.utils.RESpeckLiveData
 
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
 
 /**
  * A simple [Fragment] subclass.
@@ -34,6 +36,102 @@ private const val ARG_PARAM2 = "param2"
  */
 class DetectFragment : Fragment() {
 
+    val GENERAL_ACTIVITIES = listOf(
+        "Sitting/Standing",
+        "Lying down on left",
+        "Lying down on right",
+        "Lying down on back",
+        "Lying Down on Stomach",
+        "Walking normally",
+        "Running",
+        "Descending stairs",
+        "Ascending stairs",
+        "Shuffle walking",
+        "Miscellaneous movements"
+    )
+
+    val STATIONARY_ACTIVITIES = listOf(
+        "Sitting/standing and breathing normally",
+        "Lying down left and breathing normally",
+        "Lying down right and breathing normally",
+        "Lying down on back and breathing normally",
+        "Lying down on stomach and breathing normally",
+        "Sitting/standing and coughing",
+        "Lying down on left and coughing",
+        "Lying down on right and coughing",
+        "Lying down on back and coughing",
+        "Lying down on stomach and coughing",
+        "Sitting/standing and hyperventilating",
+        "Lying down on left and hyperventilating",
+        "Lying down on right and hyperventilating",
+        "Lying down on back and hyperventilating",
+        "Lyging down on stomach and hyperventilating"
+    )
+
+    val TASK3_ACTIVITIES = listOf(
+        "sitting_standing breathingNormal",
+        "lyingLeft breathingNormal",
+        "lyingRight breathingNormal",
+        "lyingBack breathingNormal",
+        "lyingStomach breathingNormal",
+        "sitting_standing coughing",
+        "lyingLeft coughing",
+        "lyingRight coughing",
+        "lyingBack coughing",
+        "lyingStomach coughing",
+        "sitting_standing hyperventilating",
+        "lyingLeft hyperventilating",
+        "lyingRight hyperventilating",
+        "lyingBack hyperventilating",
+        "lyingStomach hyperventilating",
+        "sitting_standing singing",
+        "lyingLeft singing",
+        "lyingRight singing",
+        "lyingBack singing",
+        "lyingStomach singing",
+        "sitting_standing laughing",
+        "lyingLeft laughing",
+        "lyingRight laughing",
+        "lyingBack laughing",
+        "lyingStomach laughing",
+        "sitting_standing talking",
+        "lyingLeft talking",
+        "lyingRight talking",
+        "lyingBack talking",
+        "lyingStomach talking",
+        "sitting/standing eating",
+    )
+
+
+    // global graph variables
+    val GENERAL_CLASSIFIER = Classifier(
+        "ten_sec.tflite",
+        50,
+        3,
+        10,
+        11,
+        GENERAL_ACTIVITIES
+    )
+
+    val STATIONARY_CLASSIFIER = Classifier(
+        "conv_model_task2_50_3.tflite",
+        50,
+        3,
+        10,
+        15,
+        STATIONARY_ACTIVITIES
+    )
+
+    val TASK3_CLASSIFIER = Classifier(
+        "cnn_model_task3_100_20.tflite",
+        100,
+        3,
+        20,
+        TASK3_ACTIVITIES.size,
+        TASK3_ACTIVITIES
+    )
+
+    var currentClassifier = GENERAL_CLASSIFIER
 
 
     lateinit var dataSet_res_accel_x: LineDataSet
@@ -58,11 +156,80 @@ class DetectFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         val view = inflater.inflate(R.layout.fragment_detect, container, false)
-        setupCharts(view)
-        setupClassifierSpinner(view)
 
-
+        detectedActivity = view.findViewById(R.id.detected_activity_text)
         return view
+    }
+
+    override fun onResume() {
+        super.onResume()
+
+        // set up the broadcast receiver
+        respeckLiveUpdateReceiver = object : BroadcastReceiver() {
+            override fun onReceive(context: Context, intent: Intent) {
+
+                Log.i("thread", "I am running on thread = " + Thread.currentThread().name)
+
+                val action = intent.action
+
+                if (action == Constants.ACTION_RESPECK_LIVE_BROADCAST) {
+
+                    val liveData =
+                        intent.getSerializableExtra(Constants.RESPECK_LIVE_DATA) as RESpeckLiveData
+                    Log.d("Live", "onReceive: liveData = $liveData")
+
+                    // get all relevant intent contents
+                    val accelX = liveData.accelX
+                    val accelY = liveData.accelY
+                    val accelZ = liveData.accelZ
+                    val gyroX = liveData.gyro.x
+                    val gyroY = liveData.gyro.y
+                    val gyroZ = liveData.gyro.z
+
+                    time += 1
+
+                    currentClassifier.addData(accelX, accelY, accelZ, gyroX, gyroY, gyroZ)
+                    if (currentClassifier.index == currentClassifier.windowSize) {
+                        val activity = currentClassifier.classifyData(context)
+                        runOnUiThread {
+                            val activityText = "Detected activity: $activity"
+                            detectedActivity.text = activityText
+                        }
+                    }
+
+                    updateGraph("respeck", accelX, accelY, accelZ)
+                }
+            }
+        }
+
+        // register receiver on another thread
+        val handlerThreadRespeck = HandlerThread("bgThreadRespeckLive")
+        handlerThreadRespeck.start()
+        looperRespeck = handlerThreadRespeck.looper
+        val handlerRespeck = Handler(looperRespeck)
+        requireContext().registerReceiver(
+            respeckLiveUpdateReceiver,
+            filterTestRespeck,
+            null,
+            handlerRespeck
+        )
+
+    }
+
+    override fun onPause() {
+        super.onPause()
+        // unregister the broadcast receiver if it's registered
+        if (this::respeckLiveUpdateReceiver.isInitialized) {
+            requireContext().unregisterReceiver(respeckLiveUpdateReceiver)
+            looperRespeck.quit()
+        }
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        setupClassifierSpinner(view)
+        setupCharts(view)
     }
 
     fun setupClassifierSpinner(view: View) {
@@ -164,113 +331,5 @@ class DetectFragment : Fragment() {
         }
     }
 
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        super.onActivityCreated(savedInstanceState)
-        GENERAL_CLASSIFIER = Classifier(
-            requireContext(),
-            "ten_sec.tflite",
-            50,
-            3,
-            10,
-            11,
-            GENERAL_ACTIVITIES
-        )
 
-        STATIONARY_CLASSIFIER = Classifier(
-            requireContext(),
-            "conv_model_task2_50_3.tflite",
-            50,
-            3,
-            10,
-            15,
-            STATIONARY_ACTIVITIES
-        )
-
-        TASK3_CLASSIFIER = Classifier(
-            requireContext(),
-            "cnn_model_task3_100_20.tflite",
-            100,
-            3,
-            20,
-            TASK3_ACTIVITIES.size,
-            TASK3_ACTIVITIES
-        )
-    }
-
-    override fun onDestroy() {
-        super.onDestroy()
-        requireContext().unregisterReceiver(respeckLiveUpdateReceiver)
-        looperRespeck.quit()
-    }
-
-    companion object {
-
-        val GENERAL_ACTIVITIES = listOf(
-            "Sitting/Standing",
-            "Lying down on left",
-            "Lying down on right",
-            "Lying down on back",
-            "Lying Down on Stomach",
-            "Walking normally",
-            "Running",
-            "Descending stairs",
-            "Ascending stairs",
-            "Shuffle walking",
-            "Miscellaneous movements"
-        )
-
-        val STATIONARY_ACTIVITIES = listOf(
-            "Sitting/standing and breathing normally",
-            "Lying down left and breathing normally",
-            "Lying down right and breathing normally",
-            "Lying down on back and breathing normally",
-            "Lying down on stomach and breathing normally",
-            "Sitting/standing and coughing",
-            "Lying down on left and coughing",
-            "Lying down on right and coughing",
-            "Lying down on back and coughing",
-            "Lying down on stomach and coughing",
-            "Sitting/standing and hyperventilating",
-            "Lying down on left and hyperventilating",
-            "Lying down on right and hyperventilating",
-            "Lying down on back and hyperventilating",
-            "Lyging down on stomach and hyperventilating"
-        )
-
-        val TASK3_ACTIVITIES = listOf(
-            "sitting_standing breathingNormal",
-            "lyingLeft breathingNormal",
-            "lyingRight breathingNormal",
-            "lyingBack breathingNormal",
-            "lyingStomach breathingNormal",
-            "sitting_standing coughing",
-            "lyingLeft coughing",
-            "lyingRight coughing",
-            "lyingBack coughing",
-            "lyingStomach coughing",
-            "sitting_standing hyperventilating",
-            "lyingLeft hyperventilating",
-            "lyingRight hyperventilating",
-            "lyingBack hyperventilating",
-            "lyingStomach hyperventilating",
-            "sitting_standing singing",
-            "lyingLeft singing",
-            "lyingRight singing",
-            "lyingBack singing",
-            "lyingStomach singing",
-            "sitting_standing laughing",
-            "lyingLeft laughing",
-            "lyingRight laughing",
-            "lyingBack laughing",
-            "lyingStomach laughing",
-            "sitting_standing talking",
-            "lyingLeft talking",
-            "lyingRight talking",
-            "lyingBack talking",
-            "lyingStomach talking",
-            "sitting/standing eating",
-        )
-
-
-    }
 }
