--- app/src/main/res/layout/activity_detect.xml
@@ -1,108 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="#023047"
-    android:textColor="#fff"
-    tools:context=".detect.DetectActivity">
-    >
-
-
-    <LinearLayout
-        android:id="@+id/respeck_graphs_detect"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        app:layout_constraintBottom_toTopOf="@+id/respeck_chart_detect"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0">
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Respeck Live Data"
-            android:textColor="#fff"
-            android:textStyle="bold" />
-
-    </LinearLayout>
-
-
-    <com.github.mikephil.charting.charts.LineChart
-        android:id="@+id/respeck_chart_detect"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="#fff"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.352"
-        tools:layout_editor_absoluteX="0dp" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/detected_activity_text"
-        app:layout_constraintVertical_bias="0.0">
-
-
-    </LinearLayout>
-
-    <TextView
-        android:id="@+id/detected_activity_text"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="1dp"
-        android:text="Detected Activity: "
-        android:textColor="#fff"
-        android:textSize="20sp"
-        android:textStyle="bold"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/respeck_chart_detect" />
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toTopOf="@+id/detected_activity_text"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/respeck_chart_detect">
-
-        <TextView
-            android:id="@+id/classfierText"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="Select Classifier:"
-            android:textColor="#fff"
-            android:textSize="20sp"
-            app:layout_constraintEnd_toStartOf="@+id/classifierSpinner"
-            app:layout_constraintHorizontal_bias="1.0"
-            app:layout_constraintStart_toStartOf="parent" />
-
-        <Spinner
-            android:id="@+id/classifierSpinner"
-            style="@style/MySpinnerStyle"
-
-            android:layout_width="200dp"
-            android:layout_height="wrap_content"
-            android:paddingVertical="5dp"
-            app:layout_constraintBottom_toTopOf="@+id/detected_activity_text"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.587"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/respeck_chart_detect"
-            app:layout_constraintVertical_bias="0.263" />
-
-    </LinearLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
